+++ src/render_settings.rs
@@ -2,7 +2,7 @@
 
 #[derive(Debug, Copy, Clone)]
 pub struct RenderSettings {
-    pub spp: usize,
+    pub spp: u32,
     pub max_depth: u32,
 }
